chore: rename types to guide-types
