+++ carrerait-roadmap/src/app/orientacion/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from "next/navigation";
 import { getGuideContent, getGuideSlugs } from "@/utils/guides";
-import { Guide } from "@/types/types";
+import { Guide } from "@/types/guide-types";
 import { CornerUpLeft } from "lucide-react";
 import Link from "next/link";
