+++ carrerait-roadmap/src/types/guide-types.ts
